+++ Code/CrowdControlMod.cs
@@ -946,6 +946,7 @@ public sealed class CrowdControlMod : Mod
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Summon));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Ranged));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Armour));
+        AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Accessory));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.HealingPotion));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Potion));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Food));