+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Added "Give Yoyo" effect (disabled).
+- Added "Give Accessory" effect (disabled).
 
 ## [2.0.11] - 2023-12-13